[edit]: Authentication using localStorage
